Move utils to their own file for re-use.
